requirements.txt
contextlib2==0.5.4
Django==1.9.7
django-auth-ldap==1.2.8
django-braces==1.9.0
django-oauth-toolkit==0.10.0
djangorestframework==3.4.1
gitdb==0.6.4
GitPython==2.0.7
ipaddr==2.1.11
MySQL-python==1.2.5
oauthlib==1.0.3
python-ldap==2.4.25
raven==5.23.0
six==1.10.0
smmap==0.9.0
